I just added more directories to organize the java codes for the GUI, this should make the project source code a bit more structured
