+++ src/main/java/org/example/gui/MainPanel/BottomPanel/BottomPanel.java
@@ -1,9 +1,11 @@
-package org.example.gui;
+package org.example.gui.MainPanel.BottomPanel;
+
+import org.example.gui.SuperPanel;
 
 import java.awt.*;
 
 public class BottomPanel extends SuperPanel {
-    BottomPanel() {
+    public BottomPanel() {
         this.setPreferredSize(new Dimension(1300, 200));
     }
 }
