+++ src/main/java/org/example/gui/MainPanel/LogoPanel/LogoPanel.java
@@ -1,9 +1,11 @@
-package org.example.gui;
+package org.example.gui.MainPanel.LogoPanel;
+
+import org.example.gui.SuperPanel;
 
 import java.awt.*;
 
 public class LogoPanel extends SuperPanel {
-    LogoPanel() {
+    public LogoPanel() {
         this.setPreferredSize(new Dimension(1300, 100));
     }
 }
